src/kafka_consumers/app/requirements.txt
aiokafka==0.8.1
psycopg2-binary==2.9.9
sshtunnel==0.4.0